Implemented "split view", which takes effect automatically when fewer than 5 streams are given by the user. This provides a better viewing experience when watching fewer streams.
When split view takes effect, the stream queue is hidden, and all streams are displayed in the main area.

* When watching 2 streams, they are displayed side-by-side with a vertical divider.
* When watching 3 streams, the primary streamer is displayed on top with 100% width. The two secondary streams are displayed below with a vertical divider.
* When watching 4 streams, they are all take up 25% of the available space, each getting a quadrant of the main space.

Activating Theater Mode will have a similar effect to the normal case - the main streamer will take up the entirety of the main space, and the chat panel width will be reduced.

In split view, like normal, the user can change the primary streamer by using the buttons at the top of the screen. This will also happen automatically.